backend/requirements.txt
# Web framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Config & validation
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0

# AI — pinned to compatible set
anthropic==0.40.0
langchain-core==0.3.28
langchain-anthropic==0.3.3
langchain-community==0.3.13
langgraph==0.2.60

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20

# Task queue
celery==5.4.0
redis==5.2.1

# Memory + RAG
chromadb==0.6.3
sentence-transformers==3.4.1
tavily-python==0.5.0

# Rate limiting
slowapi==0.1.9

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
